+++ graspit-reporters/apps/test-wdio-cucumber/features/pageobjects/login.page.js
@@ -1,6 +1,5 @@
-import { ChainablePromiseElement } from 'webdriverio';
-
-import Page from './page';
+const { $ } = require('@wdio/globals')
+const Page = require('./page');
 
 /**
  * sub page containing specific selectors and methods for a specific page
@@ -9,15 +8,15 @@ class LoginPage extends Page {
     /**
      * define selectors using getter methods
      */
-    public get inputUsername () {
+    get inputUsername () {
         return $('#username');
     }
 
-    public get inputPassword () {
+    get inputPassword () {
         return $('#password');
     }
 
-    public get btnSubmit () {
+    get btnSubmit () {
         return $('button[type="submit"]');
     }
 
@@ -25,7 +24,7 @@ class LoginPage extends Page {
      * a method to encapsule automation code to interact with the page
      * e.g. to login using username and password
      */
-    public async login (username: string, password: string) {
+    async login (username, password) {
         await this.inputUsername.setValue(username);
         await this.inputPassword.setValue(password);
         await this.btnSubmit.click();
@@ -34,9 +33,9 @@ class LoginPage extends Page {
     /**
      * overwrite specific options to adapt it to page object
      */
-    public open () {
+    open () {
         return super.open('login');
     }
 }
 
-export default new LoginPage();
+module.exports = new LoginPage();