+++ graspit-reporters/apps/test-wdio-cucumber/features/pageobjects/page.js
@@ -1,13 +1,15 @@
+const { browser } = require('@wdio/globals')
+
 /**
 * main page object containing all methods, selectors and functionality
 * that is shared across all page objects
 */
-export default class Page {
+module.exports = class Page {
     /**
     * Opens a sub page of the page
     * @param path path of the sub page (e.g. /path/to/page.html)
     */
-    public open (path: string) {
+    open (path) {
         return browser.url(`https://the-internet.herokuapp.com/${path}`)
     }
 }