+++ graspit-reporters/apps/test-wdio-cucumber/features/pageobjects/secure.page.js
@@ -1,6 +1,5 @@
-import { ChainablePromiseElement } from 'webdriverio';
-
-import Page from './page';
+const { $ } = require('@wdio/globals')
+const Page = require('./page');
 
 /**
  * sub page containing specific selectors and methods for a specific page
@@ -9,9 +8,9 @@ class SecurePage extends Page {
     /**
      * define selectors using getter methods
      */
-    public get flashAlert () {
+    get flashAlert () {
         return $('#flash');
     }
 }
 
-export default new SecurePage();
+module.exports = new SecurePage();