+++ graspit-reporters/packages/graspit-dashboard/src/components/Charts/StatusPieChart.tsx
@@ -1,112 +1,139 @@
 import type TestRunRecord from 'src/types/testRunRecords';
 import type { SuiteSummary } from 'src/types/testRunRecords';
 import React, { type ReactNode } from 'react';
-import Highcharts from 'highcharts';
-import HighchartsReact from 'highcharts-react-official';
-import highContrastDark from 'highcharts/themes/high-contrast-dark';
-import { greenGradient, redGradient, skippedGradient } from './constants';
-import { toolTipFormats } from '../utils/counter';
-import { REM } from 'next/font/google';
+import ReactECharts from 'echarts-for-react';
+import * as echarts from 'echarts/core';
+import {
+    TitleComponent,
+    TooltipComponent,
+    DatasetComponent,
+    GridComponent,
+} from 'echarts/components';
 
-const serif = REM({
-    subsets: ['latin'],
-    weight: '300',
-    adjustFontFallback: true,
-});
+import type {
+    TitleComponentOption,
+    TooltipComponentOption,
+    ToolboxComponentOption,
+    GridComponentOption,
+} from 'echarts/components';
 
-if (typeof Highcharts === 'object') {
-    highContrastDark(Highcharts);
-}
+import { PieChart } from 'echarts/charts';
+import type { PieSeriesOption } from 'echarts/charts';
+import type { ComposeOption } from 'echarts/core';
+
+// Import the Canvas renderer
+// Note that including the CanvasRenderer or SVGRenderer is a required step
+import { SVGRenderer } from 'echarts/renderers';
+import {
+    radiantGreen,
+    radiantRed,
+    radiantYellow,
+    serif,
+    toolTipFormats,
+} from './constants';
+import type { CallbackDataParams } from 'echarts/types/dist/shared';
+import type { TopLevelFormatterParams } from 'echarts/types/dist/shared';
+
+type composed = ComposeOption<
+    | PieSeriesOption
+    | GridComponentOption
+    | TitleComponentOption
+    | TooltipComponentOption
+    | ToolboxComponentOption
+    | GridComponentOption
+>;
+
+// Register the required components
+echarts.use([
+    TitleComponent,
+    SVGRenderer,
+    PieChart,
+    GridComponent,
+    TooltipComponent,
+    DatasetComponent,
+]);
 
 export default function ProgressPieChart(props: {
     run: TestRunRecord;
     isTestCases: boolean;
 }): ReactNode {
     const suite: SuiteSummary = JSON.parse(props.run.suiteSummary);
-    const total = props.isTestCases ? props.run.tests : suite.count;
+    const failed = props.isTestCases ? props.run.failed : suite.failed;
+    const passed = props.isTestCases ? props.run.passed : suite.passed;
+    const skipped = props.isTestCases ? props.run.skipped : suite.skipped;
 
-    const _overallPassed =
-        (props.isTestCases ? props.run.passed : suite.passed) / total;
-    const passPercent = Number.isNaN(_overallPassed) ? 0 : _overallPassed;
-
-    const options: Highcharts.Options = {
-        credits: { enabled: false, text: 'link-to-docs-in-future' },
-        exporting: { enabled: false },
-
-        chart: {
-            type: 'pie',
-            backgroundColor: 'transparent',
-            shadow: true,
-            spacing: [0, 0, 0, 0],
-            margin: [0, 0, 0, 0],
-            height: 120,
-            style: { fontFamily: serif.style.fontFamily },
-        },
-        tooltip: toolTipFormats,
-        title: {
-            text: `${(passPercent * 100).toFixed(2)}%`,
-            x: -20,
-            y: 40,
-            align: 'center',
-            verticalAlign: 'middle',
-            style: {
-                fontSize: '12',
-                color: 'white',
+    const options: composed = {
+        tooltip: {
+            trigger: 'item',
+            formatter: (param: TopLevelFormatterParams) => {
+                const args = param as CallbackDataParams;
+                // correct the percentage
+                return `${args.seriesName}<br/>${args.marker}  ${args.name}: ${
+                    args.value
+                } (${(args.percent ?? 0) * 2}%)`;
             },
+            ...toolTipFormats,
         },
-        subtitle: {
-            text: undefined,
-        },
-        plotOptions: {
-            pie: {
-                shadow: {
-                    offsetX: 0,
-                    offsetY: 5,
-                    opacity: 0.2,
-                    width: 5,
-                },
-                innerSize: 90,
-                colors: [greenGradient, redGradient, skippedGradient],
-                startAngle: -90,
-                endAngle: 90,
-                center: ['43%', '100%'],
-                size: '180%',
-                borderWidth: 0,
-            },
+        textStyle: {
+            fontFamily: serif.style.fontFamily,
         },
         series: [
             {
-                allowPointSelect: false,
-                shadow: true,
-
+                name: props.isTestCases ? 'Tests' : 'Suites',
                 type: 'pie',
-                name: !props.isTestCases ? 'Suite Entities' : 'Test Entities',
-                data: [
-                    [
-                        'Passed',
-                        props.isTestCases ? props.run.passed : suite.passed,
-                    ],
-                    [
-                        'Failed',
-                        props.isTestCases ? props.run.failed : suite.failed,
-                    ],
-                    [
-                        'Skipped',
-                        props.isTestCases ? props.run.skipped : suite.skipped,
-                    ],
-                ],
-                dataLabels: {
-                    style: {
-                        textOutline: '0px',
-                        color: 'white',
-                        fontWeight: 'normal',
-                    },
-                    alignTo: 'center',
-                    distance: 10,
+                top: 80,
+                left: -35,
+                radius: ['200%', '400%'],
+                startAngle: 180,
+                itemStyle: {
+                    borderRadius: 5,
                 },
+                label: {
+                    show: true,
+                    color: 'white',
+                    verticalAlign: 'bottom',
+                    formatter: '{b}:{c}',
+                },
+                data: [
+                    {
+                        value: passed,
+                        name: 'Passed',
+                        itemStyle: { color: radiantGreen },
+                    },
+                    {
+                        value: failed,
+                        name: 'Failed',
+                        itemStyle: { color: radiantRed },
+                    },
+                    {
+                        value: skipped,
+                        name: 'Skipped',
+                        itemStyle: { color: radiantYellow },
+                    },
+                    {
+                        name: 'Buffer',
+                        itemStyle: {
+                            // stop the chart from rendering this piece
+                            color: 'none',
+                            decal: {
+                                symbol: 'none',
+                            },
+                        },
+                        label: {
+                            show: false,
+                        },
+                        value: passed + skipped + failed,
+                    },
+                ],
             },
         ],
     };
 
-    return <HighchartsReact highcharts={Highcharts} options={options} />;
+    return (
+        <ReactECharts
+            option={options}
+            style={{ height: '120px' }}
+            opts={{ renderer: 'svg' }}
+        />
+    );
 }