+++ graspit-reporters/packages/graspit-dashboard/src/types/uiConstants.ts
@@ -3,4 +3,4 @@ export const treeViewMode = 'tree';
 
 export const testEntitiesTab = 'test-entities';
 export const overviewTab = 'overview';
-export const ganttChartTab = 'ganttChart';
+export const timelineTab = 'timelineTab';