+++ graspit-reporters/packages/wdio-graspit-reporter/src/helpers.ts
@@ -23,15 +23,18 @@ export function attachReporter(
     },
   ]);
 
+
   toModify.services.push([
     GraspItService, {
       port,
       exePath: options.exePath,
-      projectName: options.projectName,
       timeout: options.timeout,
       root: options.root,
       collectionName: options.collectionName,
       export: options.export,
+      testConfig: {
+        ...options.testConfig, avoidParentSuitesInCount: options.testConfig.avoidParentSuitesInCount ?? config.framework === 'cucumber'
+      },
     },
   ]);