+++ graspit/services/DBService/models/config_base.py
@@ -7,7 +7,6 @@ from tortoise.fields import (
     CharField,
     TextField,
     UUIDField,
-    BooleanField,
 )