+++ graspit-reporters/packages/graspit-dashboard/CHANGELOG.md
@@ -1,5 +1,11 @@
 # graspit
 
+## 3.3.0
+
+### Minor Changes
+
+- feat: migrated to echarts
+
 ## 3.2.0
 
 ### Minor Changes