+++ graspit-reporters/packages/wdio-graspit-reporter/CHANGELOG.md
@@ -1,5 +1,16 @@
 # wdio-graspit-reporter
 
+## 3.3.0
+
+### Minor Changes
+
+- feat: migrated to echarts
+
+### Patch Changes
+
+- Updated dependencies
+  - graspit-commons@3.3.0
+
 ## 3.2.0
 
 ### Minor Changes